finished the naive scoring with NN-smoothing
